Task and Database classes added

+++ app/src/main/java/com/single/point/core/data/database/TaskDao.kt
@@ -7,7 +7,7 @@ import androidx.room.Insert
 @Dao
 interface TaskDao {
     @Insert
-    suspend fun insertTask(task:Task):Integer
+    suspend fun insertTask(task:Task):Long
 
 
 }
